Adaptations for 16:9 resolution ratio
Background mov player videoplane changed to ‘@transform_reset 2’ to
fill the whole screen. This means users will need to take care to use
correct resolution backgrounds for their projector/screen settings.
Also did this on webcam input, but there is no control to choose if it
is a 4:3 or 16:9 input. Really should sense the camera input resolution
and set accordingly.